Implementierung von Code- und Entwicklerstatistiken-Collector-Klassen sowie Git-Statistiken- und Qualitätsmetriken-Berechnung. Einführung von Modellen für Code-, Entwickler-, Git- und Qualitätsmetriken. Verbesserung der Berichtsgenerierung für Repository-Analysen.
