+++ src/analytics/CodeAnalytics.cs
@@ -1,5 +1,7 @@
 using System.Text.Json;
 using System.Text.RegularExpressions;
+using WaterWizard.Analytics.Models;
+using WaterWizard.Analytics.Core;
 
 namespace WaterWizard.Analytics;
 
@@ -9,9 +11,6 @@ namespace WaterWizard.Analytics;
 /// </summary>
 public static class CodeAnalytics
 {
-    private static readonly string[] CodeExtensions = { ".cs", ".csproj", ".sln", ".xml", ".json", ".md", ".yml", ".yaml" };
-    private static readonly string[] IgnorePatterns = { "bin/", "obj/", ".git/", "node_modules/", ".vs/", ".idea/" };
-    
     // Dateien die für Analytics nicht benötigt werden
     private static readonly string[] AnalyticsIgnoreFiles = { 
         ".DS_Store", 
@@ -24,7 +23,7 @@ public static class CodeAnalytics
         "*.swo",
         "*~"
     };
-    
+
     /// <summary>
     /// Hauptmethode für die Generierung von Repository-Statistiken
     /// </summary>
@@ -42,16 +41,20 @@ public static class CodeAnalytics
         try
         {
             // Git-Statistiken sammeln
-            await CollectGitStatistics(repositoryPath, analytics);
+            var gitCollector = new GitStatisticsCollector(repositoryPath);
+            analytics.GitStatistics = await gitCollector.CollectAsync();
             
             // Code-Statistiken sammeln
-            CollectCodeStatistics(repositoryPath, analytics);
+            var codeCollector = new CodeStatisticsCollector(repositoryPath);
+            analytics.CodeStatistics = codeCollector.Collect();
             
             // Entwickler-Statistiken sammeln
-            await CollectDeveloperStatistics(repositoryPath, analytics);
+            var devCollector = new DeveloperStatisticsCollector(repositoryPath);
+            analytics.DeveloperStatistics = await devCollector.CollectAsync();
             
             // Qualitäts-Metriken berechnen
-            CalculateQualityMetrics(analytics);
+            var metricsCalculator = new QualityMetricsCalculator(analytics);
+            analytics.QualityMetrics = metricsCalculator.Calculate();
             
             // Report generieren
             var report = GenerateReport(analytics);
@@ -73,362 +76,6 @@ public static class CodeAnalytics
             Console.WriteLine($"[CodeAnalytics] Fehler bei der Analyse: {ex.Message}");
             return JsonSerializer.Serialize(new { Error = ex.Message, GeneratedAt = DateTime.UtcNow });
         }
-    }
-    
-    /// <summary>
-    /// Sammelt Git-Statistiken über Commits, Branches und Tags
-    /// </summary>
-    private static async Task CollectGitStatistics(string repositoryPath, RepositoryAnalytics analytics)
-    {
-        try
-        {
-            // Git-Status abrufen
-            var gitStatus = await ExecuteGitCommand(repositoryPath, "status --porcelain");
-            analytics.GitStatistics = new GitStatistics
-            {
-                UncommittedChanges = gitStatus.Split('\n', StringSplitOptions.RemoveEmptyEntries).Length,
-                CurrentBranch = await ExecuteGitCommand(repositoryPath, "branch --show-current"),
-                TotalCommits = int.Parse(await ExecuteGitCommand(repositoryPath, "rev-list --count HEAD")),
-                TotalBranches = int.Parse(await ExecuteGitCommand(repositoryPath, "branch -r | wc -l")),
-                LastCommit = await ExecuteGitCommand(repositoryPath, "log -1 --format=%H"),
-                LastCommitMessage = await ExecuteGitCommand(repositoryPath, "log -1 --format=%s"),
-                LastCommitAuthor = await ExecuteGitCommand(repositoryPath, "log -1 --format=%an"),
-                LastCommitDate = DateTime.Parse(await ExecuteGitCommand(repositoryPath, "log -1 --format=%ci"))
-            };
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"[CodeAnalytics] Git-Statistiken konnten nicht gesammelt werden: {ex.Message}");
-        }
-    }
-    
-    /// <summary>
-    /// Sammelt Code-Statistiken über Dateien, Zeilen und Komplexität
-    /// </summary>
-    private static void CollectCodeStatistics(string repositoryPath, RepositoryAnalytics analytics)
-    {
-        var codeStats = new CodeStatistics();
-        var allFiles = GetAllCodeFiles(repositoryPath);
-        
-        foreach (var file in allFiles)
-        {
-            try
-            {
-                var fileInfo = AnalyzeFile(file);
-                codeStats.TotalFiles++;
-                codeStats.TotalLines += fileInfo.TotalLines;
-                codeStats.CodeLines += fileInfo.CodeLines;
-                codeStats.CommentLines += fileInfo.CommentLines;
-                codeStats.EmptyLines += fileInfo.EmptyLines;
-                codeStats.TotalSize += fileInfo.FileSize;
-                
-                // Dateityp-spezifische Statistiken
-                var extension = Path.GetExtension(file).ToLower();
-                if (!codeStats.FilesByType.ContainsKey(extension))
-                    codeStats.FilesByType[extension] = 0;
-                codeStats.FilesByType[extension]++;
-                
-                // Komplexitäts-Metriken
-                if (extension == ".cs")
-                {
-                    codeStats.CSharpFiles++;
-                    codeStats.Classes += CountPattern(file, @"\bclass\s+\w+");
-                    codeStats.Methods += CountPattern(file, @"\b(public|private|protected|internal)\s+\w+\s+\w+\s*\([^)]*\)\s*\{");
-                    codeStats.Interfaces += CountPattern(file, @"\binterface\s+\w+");
-                    codeStats.Properties += CountPattern(file, @"\b(public|private|protected|internal)\s+\w+\s+\w+\s*\{\s*(get|set)");
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"[CodeAnalytics] Fehler bei der Analyse von {file}: {ex.Message}");
-            }
-        }
-        
-        analytics.CodeStatistics = codeStats;
-    }
-    
-    /// <summary>
-    /// Sammelt Entwickler-Statistiken über Commits und Aktivität
-    /// </summary>
-    private static async Task CollectDeveloperStatistics(string repositoryPath, RepositoryAnalytics analytics)
-    {
-        try
-        {
-            // Mappe alle Originalnamen auf den normierten Namen
-            var nameToNormalized = new Dictionary<string, string>();
-            var normalizedToAliases = new Dictionary<string, HashSet<string>>();
-            var developerStats = new Dictionary<string, DeveloperStatistics>();
-            
-            // Git-Log für Entwickler-Statistiken
-            var gitLog = await ExecuteGitCommand(repositoryPath, "log --pretty=format:\"%an|%ad|%s\" --date=short");
-            var commits = gitLog.Split('\n', StringSplitOptions.RemoveEmptyEntries);
-            
-            foreach (var commit in commits)
-            {
-                var parts = commit.Split('|');
-                if (parts.Length >= 3)
-                {
-                    var author = parts[0];
-                    var dateStr = parts[1];
-                    var message = parts[2];
-                    
-                    // Korrektes Datum parsen
-                    if (!DateTime.TryParse(dateStr, out var date))
-                    {
-                        Console.WriteLine($"[CodeAnalytics] Konnte Datum nicht parsen: {dateStr}");
-                        continue;
-                    }
-                    
-                    // Normalisiere den Autorennamen für bessere Gruppierung
-                    var normalizedAuthor = NormalizeAuthorName(author);
-                    nameToNormalized[author] = normalizedAuthor;
-                    if (!normalizedToAliases.ContainsKey(normalizedAuthor))
-                        normalizedToAliases[normalizedAuthor] = new HashSet<string>();
-                    normalizedToAliases[normalizedAuthor].Add(author);
-                    
-                    if (!developerStats.ContainsKey(normalizedAuthor))
-                    {
-                        developerStats[normalizedAuthor] = new DeveloperStatistics
-                        {
-                            Name = normalizedAuthor,
-                            OriginalNames = new List<string>(),
-                            FirstCommit = date,
-                            LastCommit = date,
-                            TotalCommits = 0,
-                            CommitMessages = new List<string>(),
-                            WeeklyActivity = new Dictionary<string, int>(),
-                            MonthlyActivity = new Dictionary<string, int>()
-                        };
-                    }
-                    
-                    var dev = developerStats[normalizedAuthor];
-                    dev.TotalCommits++;
-                    
-                    // Korrekte First/Last Commit Logik
-                    if (date < dev.FirstCommit)
-                        dev.FirstCommit = date;
-                    if (date > dev.LastCommit)
-                        dev.LastCommit = date;
-                    
-                    dev.CommitMessages.Add(message);
-                    
-                    // Wöchentliche und monatliche Aktivität mit korrektem ISO-Format
-                    var weekKey = GetIsoWeekKey(date);
-                    var monthKey = date.ToString("yyyy-MM");
-                    
-                    if (!dev.WeeklyActivity.ContainsKey(weekKey))
-                        dev.WeeklyActivity[weekKey] = 0;
-                    if (!dev.MonthlyActivity.ContainsKey(monthKey))
-                        dev.MonthlyActivity[monthKey] = 0;
-                    
-                    dev.WeeklyActivity[weekKey]++;
-                    dev.MonthlyActivity[monthKey]++;
-                    
-                    // Commit-Typ-Analyse basierend auf WaterWizards-Konventionen
-                    if (IsFeatureCommit(message))
-                        dev.FeatureCommits++;
-                    else if (IsBugFixCommit(message))
-                        dev.BugFixCommits++;
-                    else if (message.StartsWith("refactor:", StringComparison.OrdinalIgnoreCase))
-                        dev.RefactorCommits++;
-                    else if (message.StartsWith("docs:", StringComparison.OrdinalIgnoreCase))
-                        dev.DocumentationCommits++;
-                }
-            }
-            // Aliase zuordnen (alle Originalnamen, die auf denselben normierten Namen mappen)
-            foreach (var kv in developerStats)
-            {
-                if (normalizedToAliases.ContainsKey(kv.Key))
-                    kv.Value.OriginalNames = normalizedToAliases[kv.Key].OrderBy(x => x).ToList();
-            }
-            analytics.DeveloperStatistics = developerStats.Values.ToList();
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"[CodeAnalytics] Entwickler-Statistiken konnten nicht gesammelt werden: {ex.Message}");
-        }
-    }
-    
-    /// <summary>
-    /// Normalisiert Autorennamen für bessere Gruppierung ähnlicher Namen
-    /// </summary>
-    private static string NormalizeAuthorName(string authorName)
-    {
-        var normalized = authorName.ToLowerInvariant().Trim();
-        
-        // Alle Varianten von Erick auf exakt "Erick Zeiler" mappen
-        if (normalized == "erick" || normalized == "erickk0" || normalized == "erick zeiler")
-            return "Erick Zeiler";
-        
-        // Spezielle Fälle für bekannte Entwickler
-        var nameMappings = new Dictionary<string, string>
-        {
-            { "justin", "Justin Dewitz" },
-            { "justinjd00", "Justin Dewitz" },
-            { "justindew", "Justin Dewitz" },
-            { "jdewi001", "Justin Dewitz" },
-            { "max", "Max Kondratov" },
-            { "maxk2807", "Max Kondratov" },
-            { "maxkondratov", "Max Kondratov" },
-            { "julian", "Julian" },
-            { "jlnhsrm", "Julian" },
-            { "paul", "Paul" }
-        };
-        if (nameMappings.ContainsKey(normalized))
-            return nameMappings[normalized];
-        return char.ToUpperInvariant(authorName[0]) + authorName.Substring(1).ToLowerInvariant();
-    }
-    
-    /// <summary>
-    /// Prüft ob ein Commit ein Feature-Commit ist (F{TicketNumber} Konvention)
-    /// </summary>
-    private static bool IsFeatureCommit(string commitMessage)
-    {
-        // WaterWizards Feature-Konvention: F{TicketNumber}
-        if (Regex.IsMatch(commitMessage, @"\bF\d+\b", RegexOptions.IgnoreCase))
-            return true;
-            
-        // Standard-Konventionen als Fallback
-        if (commitMessage.StartsWith("feat:", StringComparison.OrdinalIgnoreCase))
-            return true;
-            
-        // Feature-bezogene Keywords
-        var featureKeywords = new[] { "feature", "add", "implement", "new", "create" };
-        return featureKeywords.Any(keyword => 
-            commitMessage.Contains(keyword, StringComparison.OrdinalIgnoreCase));
-    }
-    
-    /// <summary>
-    /// Prüft ob ein Commit ein Bugfix-Commit ist (B{TicketNumber} Konvention)
-    /// </summary>
-    private static bool IsBugFixCommit(string commitMessage)
-    {
-        // WaterWizards Bug-Konvention: B{TicketNumber}
-        if (Regex.IsMatch(commitMessage, @"\bB\d+\b", RegexOptions.IgnoreCase))
-            return true;
-            
-        // Standard-Konventionen als Fallback
-        if (commitMessage.StartsWith("fix:", StringComparison.OrdinalIgnoreCase))
-            return true;
-            
-        // Bugfix-bezogene Keywords
-        var bugKeywords = new[] { "bug", "fix", "repair", "resolve", "correct", "patch" };
-        return bugKeywords.Any(keyword => 
-            commitMessage.Contains(keyword, StringComparison.OrdinalIgnoreCase));
-    }
-    
-    /// <summary>
-    /// Berechnet Qualitäts-Metriken für das Projekt
-    /// </summary>
-    private static void CalculateQualityMetrics(RepositoryAnalytics analytics)
-    {
-        var qualityMetrics = new QualityMetrics();
-        
-        if (analytics.CodeStatistics != null)
-        {
-            var stats = analytics.CodeStatistics;
-            
-            // Code-Qualitäts-Metriken
-            qualityMetrics.CodeToCommentRatio = stats.CommentLines > 0 ? (double)stats.CodeLines / stats.CommentLines : 0;
-            qualityMetrics.EmptyLinesPercentage = stats.TotalLines > 0 ? (double)stats.EmptyLines / stats.TotalLines * 100 : 0;
-            qualityMetrics.AverageFileSize = stats.TotalFiles > 0 ? stats.TotalSize / stats.TotalFiles : 0;
-            qualityMetrics.AverageLinesPerFile = stats.TotalFiles > 0 ? stats.TotalLines / stats.TotalFiles : 0;
-            
-            // Komplexitäts-Metriken
-            if (stats.CSharpFiles > 0)
-            {
-                qualityMetrics.AverageMethodsPerClass = stats.Classes > 0 ? (double)stats.Methods / stats.Classes : 0;
-                qualityMetrics.AveragePropertiesPerClass = stats.Classes > 0 ? (double)stats.Properties / stats.Classes : 0;
-                qualityMetrics.CodeComplexity = stats.Methods > 0 ? (double)stats.CodeLines / stats.Methods : 0;
-                qualityMetrics.DocumentationCoverage = stats.Classes > 0 ? (double)stats.CommentLines / stats.Classes : 0;
-            }
-        }
-        
-        // Entwickler-Aktivitäts-Metriken
-        if (analytics.DeveloperStatistics != null && analytics.DeveloperStatistics.Any())
-        {
-            qualityMetrics.TotalDevelopers = analytics.DeveloperStatistics.Count;
-            
-            // Erweiterte Entwickler-Statistiken berechnen
-            foreach (var dev in analytics.DeveloperStatistics)
-            {
-                // Commit-Percentages
-                dev.FeatureCommitPercentage = dev.TotalCommits > 0 ? (double)dev.FeatureCommits / dev.TotalCommits * 100 : 0;
-                dev.BugFixCommitPercentage = dev.TotalCommits > 0 ? (double)dev.BugFixCommits / dev.TotalCommits * 100 : 0;
-                dev.RefactorCommitPercentage = dev.TotalCommits > 0 ? (double)dev.RefactorCommits / dev.TotalCommits * 100 : 0;
-                dev.DocumentationCommitPercentage = dev.TotalCommits > 0 ? (double)dev.DocumentationCommits / dev.TotalCommits * 100 : 0;
-                
-                // Zeit-basierte Statistiken
-                var now = DateTime.Now;
-                dev.DaysSinceFirstCommit = (now - dev.FirstCommit).Days;
-                dev.DaysSinceLastCommit = (now - dev.LastCommit).Days;
-                dev.CommitFrequency = dev.DaysSinceFirstCommit > 0 ? (double)dev.TotalCommits / dev.DaysSinceFirstCommit : 0;
-                
-                // Wöchentliche und monatliche Durchschnitte
-                dev.AverageCommitsPerWeek = dev.WeeklyActivity.Count > 0 ? dev.WeeklyActivity.Values.Average() : 0;
-                dev.AverageCommitsPerMonth = dev.MonthlyActivity.Count > 0 ? dev.MonthlyActivity.Values.Average() : 0;
-                
-                // Aktivste Perioden
-                dev.MostActiveWeek = dev.WeeklyActivity.Count > 0 ? 
-                    dev.WeeklyActivity.OrderByDescending(x => x.Value).First().Key : "";
-                dev.MostActiveMonth = dev.MonthlyActivity.Count > 0 ? 
-                    dev.MonthlyActivity.OrderByDescending(x => x.Value).First().Key : "";
-                dev.MaxCommitsInWeek = dev.WeeklyActivity.Count > 0 ? dev.WeeklyActivity.Values.Max() : 0;
-                dev.MaxCommitsInMonth = dev.MonthlyActivity.Count > 0 ? dev.MonthlyActivity.Values.Max() : 0;
-            }
-            
-            // Verbesserte Most Active Developer Berechnung
-            // Gewichtete Bewertung: Commits (30%) + Features (25%) + Bugfixes (20%) + Aktivität (15%) + Dokumentation (10%)
-            var developerScores = analytics.DeveloperStatistics.Select(dev => new
-            {
-                Developer = dev,
-                Score = (dev.TotalCommits * 0.3) + 
-                       (dev.FeatureCommits * 0.25) + 
-                       (dev.BugFixCommits * 0.2) + 
-                       (dev.AverageCommitsPerWeek * 0.15) + 
-                       (dev.DocumentationCommits * 0.1)
-            }).OrderByDescending(x => x.Score).ToList();
-            
-            qualityMetrics.MostActiveDeveloper = developerScores.FirstOrDefault()?.Developer.Name ?? "Unknown";
-            qualityMetrics.AverageCommitsPerDeveloper = analytics.DeveloperStatistics.Average(d => d.TotalCommits);
-            
-            // Zusätzliche Metriken für bessere Bewertung
-            qualityMetrics.TopDeveloperScore = developerScores.FirstOrDefault()?.Score ?? 0;
-            qualityMetrics.TopDeveloperBreakdown = developerScores.FirstOrDefault()?.Developer != null ? 
-                $"Commits: {developerScores.First().Developer.TotalCommits}, " +
-                $"Features: {developerScores.First().Developer.FeatureCommits}, " +
-                $"Bugfixes: {developerScores.First().Developer.BugFixCommits}, " +
-                $"Weekly Avg: {developerScores.First().Developer.AverageCommitsPerWeek:F1}, " +
-                $"Docs: {developerScores.First().Developer.DocumentationCommits}" : "";
-            
-            // Projekt-weite Statistiken
-            var allWeeks = analytics.DeveloperStatistics
-                .SelectMany(d => d.WeeklyActivity)
-                .GroupBy(x => x.Key)
-                .ToDictionary(g => g.Key, g => g.Sum(x => x.Value));
-            
-            var allMonths = analytics.DeveloperStatistics
-                .SelectMany(d => d.MonthlyActivity)
-                .GroupBy(x => x.Key)
-                .ToDictionary(g => g.Key, g => g.Sum(x => x.Value));
-            
-            qualityMetrics.MostActiveWeek = allWeeks.Count > 0 ? 
-                allWeeks.OrderByDescending(x => x.Value).First().Key : "";
-            qualityMetrics.MostActiveMonth = allMonths.Count > 0 ? 
-                allMonths.OrderByDescending(x => x.Value).First().Key : "";
-            
-            // Repository-Alter berechnen
-            var firstCommit = analytics.DeveloperStatistics.Min(d => d.FirstCommit);
-            var lastCommit = analytics.DeveloperStatistics.Max(d => d.LastCommit);
-            var repositoryAge = (lastCommit - firstCommit).Days;
-            qualityMetrics.RepositoryAge = $"{repositoryAge} days ({repositoryAge / 365.25:F1} years)";
-            
-            // Projekt-Velocity (Commits pro Tag im Durchschnitt)
-            qualityMetrics.ProjectVelocity = repositoryAge > 0 ? 
-                (double)analytics.DeveloperStatistics.Sum(d => d.TotalCommits) / repositoryAge : 0;
-        }
-        
-        analytics.QualityMetrics = qualityMetrics;
     }
     
     /// <summary>
@@ -677,258 +324,4 @@ public static class CodeAnalytics
             Console.WriteLine($"[CodeAnalytics] Fehler beim Aufräumen: {ex.Message}");
         }
     }
-    
-    #region Helper Methods
-    
-    private static async Task<string> ExecuteGitCommand(string repositoryPath, string command)
-    {
-        var process = new System.Diagnostics.Process
-        {
-            StartInfo = new System.Diagnostics.ProcessStartInfo
-            {
-                FileName = "git",
-                Arguments = command,
-                WorkingDirectory = repositoryPath,
-                RedirectStandardOutput = true,
-                RedirectStandardError = true,
-                UseShellExecute = false,
-                CreateNoWindow = true
-            }
-        };
-        
-        process.Start();
-        var output = await process.StandardOutput.ReadToEndAsync();
-        await process.WaitForExitAsync();
-        
-        return output.Trim();
-    }
-    
-    private static IEnumerable<string> GetAllCodeFiles(string repositoryPath)
-    {
-        var files = new List<string>();
-        
-        foreach (var file in Directory.GetFiles(repositoryPath, "*.*", SearchOption.AllDirectories))
-        {
-            var relativePath = Path.GetRelativePath(repositoryPath, file);
-            var fileName = Path.GetFileName(file);
-            
-            // Ignoriere bestimmte Verzeichnisse
-            if (IgnorePatterns.Any(pattern => relativePath.Contains(pattern)))
-                continue;
-                
-            // Ignoriere unnötige Dateien
-            if (AnalyticsIgnoreFiles.Any(pattern => 
-                fileName.Equals(pattern, StringComparison.OrdinalIgnoreCase) ||
-                (pattern.StartsWith("*.") && fileName.EndsWith(pattern.Substring(1), StringComparison.OrdinalIgnoreCase))))
-                continue;
-            
-            // Nur Code-Dateien
-            if (CodeExtensions.Contains(Path.GetExtension(file).ToLower()))
-            {
-                files.Add(file);
-            }
-        }
-        
-        return files;
-    }
-    
-    private static FileAnalysis AnalyzeFile(string filePath)
-    {
-        var lines = File.ReadAllLines(filePath);
-        var analysis = new FileAnalysis
-        {
-            FileSize = new FileInfo(filePath).Length,
-            TotalLines = lines.Length,
-            CodeLines = 0,
-            CommentLines = 0,
-            EmptyLines = 0
-        };
-        
-        bool inMultiLineComment = false;
-        
-        foreach (var line in lines)
-        {
-            var trimmedLine = line.Trim();
-            
-            if (string.IsNullOrEmpty(trimmedLine))
-            {
-                analysis.EmptyLines++;
-            }
-            else if (trimmedLine.StartsWith("//") || trimmedLine.StartsWith("///"))
-            {
-                analysis.CommentLines++;
-            }
-            else if (trimmedLine.StartsWith("/*"))
-            {
-                analysis.CommentLines++;
-                inMultiLineComment = true;
-            }
-            else if (trimmedLine.Contains("*/"))
-            {
-                analysis.CommentLines++;
-                inMultiLineComment = false;
-            }
-            else if (inMultiLineComment)
-            {
-                analysis.CommentLines++;
-            }
-            else
-            {
-                analysis.CodeLines++;
-            }
-        }
-        
-        return analysis;
-    }
-    
-    private static int CountPattern(string filePath, string pattern)
-    {
-        try
-        {
-            var content = File.ReadAllText(filePath);
-            return Regex.Matches(content, pattern, RegexOptions.Multiline).Count;
-        }
-        catch
-        {
-            return 0;
-        }
-    }
-    
-    /// <summary>
-    /// Berechnet den ISO-Wochen-Schlüssel für ein Datum
-    /// </summary>
-    private static string GetIsoWeekKey(DateTime date)
-    {
-        // Einfache ISO-Woche-Berechnung
-        var calendar = System.Globalization.CultureInfo.InvariantCulture.Calendar;
-        var weekOfYear = calendar.GetWeekOfYear(date, System.Globalization.CalendarWeekRule.FirstFourDayWeek, DayOfWeek.Monday);
-        var year = date.Year;
-        
-        // Wenn wir in der letzten Woche des Jahres sind, aber die Woche gehört zum nächsten Jahr
-        if (date.Month == 12 && weekOfYear > 50)
-        {
-            year++;
-            weekOfYear = 1;
-        }
-        
-        return $"{year}-W{weekOfYear:D2}";
-    }
-    
-    #endregion
-}
-
-#region Data Models
-
-public class RepositoryAnalytics
-{
-    public DateTime GeneratedAt { get; set; }
-    public string RepositoryPath { get; set; } = "";
-    public string ProjectName { get; set; } = "";
-    public GitStatistics? GitStatistics { get; set; }
-    public CodeStatistics? CodeStatistics { get; set; }
-    public List<DeveloperStatistics>? DeveloperStatistics { get; set; }
-    public QualityMetrics? QualityMetrics { get; set; }
-}
-
-public class GitStatistics
-{
-    public int UncommittedChanges { get; set; }
-    public string CurrentBranch { get; set; } = "";
-    public int TotalCommits { get; set; }
-    public int TotalBranches { get; set; }
-    public string LastCommit { get; set; } = "";
-    public string LastCommitMessage { get; set; } = "";
-    public string LastCommitAuthor { get; set; } = "";
-    public DateTime LastCommitDate { get; set; }
-}
-
-public class CodeStatistics
-{
-    public int TotalFiles { get; set; }
-    public int TotalLines { get; set; }
-    public int CodeLines { get; set; }
-    public int CommentLines { get; set; }
-    public int EmptyLines { get; set; }
-    public long TotalSize { get; set; }
-    public Dictionary<string, int> FilesByType { get; set; } = new();
-    public int CSharpFiles { get; set; }
-    public int Classes { get; set; }
-    public int Methods { get; set; }
-    public int Interfaces { get; set; }
-    public int Properties { get; set; }
-}
-
-public class DeveloperStatistics
-{
-    public string Name { get; set; } = "";
-    public DateTime FirstCommit { get; set; }
-    public DateTime LastCommit { get; set; }
-    public int TotalCommits { get; set; }
-    public int FeatureCommits { get; set; }
-    public int BugFixCommits { get; set; }
-    public int RefactorCommits { get; set; }
-    public int DocumentationCommits { get; set; }
-    public List<string> CommitMessages { get; set; } = new();
-    public List<string> OriginalNames { get; set; } = new();
-    public Dictionary<string, int> WeeklyActivity { get; set; } = new();
-    public Dictionary<string, int> MonthlyActivity { get; set; } = new();
-    
-    // Erweiterte Statistiken
-    public double AverageCommitsPerWeek { get; set; }
-    public double AverageCommitsPerMonth { get; set; }
-    public string MostActiveWeek { get; set; } = "";
-    public string MostActiveMonth { get; set; } = "";
-    public int MaxCommitsInWeek { get; set; }
-    public int MaxCommitsInMonth { get; set; }
-    public double FeatureCommitPercentage { get; set; }
-    public double BugFixCommitPercentage { get; set; }
-    public double RefactorCommitPercentage { get; set; }
-    public double DocumentationCommitPercentage { get; set; }
-    public int DaysSinceFirstCommit { get; set; }
-    public int DaysSinceLastCommit { get; set; }
-    public double CommitFrequency { get; set; } // Commits pro Tag
-}
-
-public class QualityMetrics
-{
-    public double CodeToCommentRatio { get; set; }
-    public double EmptyLinesPercentage { get; set; }
-    public double AverageFileSize { get; set; }
-    public double AverageLinesPerFile { get; set; }
-    public double AverageMethodsPerClass { get; set; }
-    public double AveragePropertiesPerClass { get; set; }
-    public int TotalDevelopers { get; set; }
-    public string MostActiveDeveloper { get; set; } = "";
-    public double AverageCommitsPerDeveloper { get; set; }
-    public double TopDeveloperScore { get; set; }
-    public string TopDeveloperBreakdown { get; set; } = "";
-    
-    // Erweiterte Metriken
-    public int TotalPullRequests { get; set; }
-    public int MergedPullRequests { get; set; }
-    public double AverageLinesPerCommit { get; set; }
-    public string MostActiveWeek { get; set; } = "";
-    public string MostActiveMonth { get; set; } = "";
-    public int MaxCommitsInDay { get; set; }
-    public double ProjectVelocity { get; set; } // Commits pro Tag im Durchschnitt
-    public double CodeComplexity { get; set; } // Durchschnittliche Methodenlänge
-    public double TestCoverage { get; set; } // Platzhalter für Test-Coverage
-    public double DocumentationCoverage { get; set; } // Anteil dokumentierter Klassen
-    public string RepositoryAge { get; set; } = "";
-    public int TotalIssues { get; set; }
-    public int ClosedIssues { get; set; }
-    public double IssueResolutionRate { get; set; }
-    public string MostProductiveDay { get; set; } = "";
-    public string MostProductiveHour { get; set; } = "";
-}
-
-public class FileAnalysis
-{
-    public long FileSize { get; set; }
-    public int TotalLines { get; set; }
-    public int CodeLines { get; set; }
-    public int CommentLines { get; set; }
-    public int EmptyLines { get; set; }
-}
-
-#endregion 
+}
