+++ src/analytics/Models/FileAnalysis.cs
@@ -1,0 +1,10 @@
+namespace WaterWizard.Analytics.Models;
+
+public class FileAnalysis
+{
+    public long FileSize { get; set; }
+    public int TotalLines { get; set; }
+    public int CodeLines { get; set; }
+    public int CommentLines { get; set; }
+    public int EmptyLines { get; set; }
+}
